requirements.txt
mysql-connector-python
fastapi
uvicorn
pydantic
pandas
numpy
transformers
beautifulsoup4
openai
python-dotenv


